ai-service/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
numpy
joblib
pydantic-settings
